fix bugs

+++ A2/getIntraPredictionReference.m
@@ -1,6 +1,6 @@
 function [verticalRefernce, horizontalReference] = getIntraPredictionReference(heightBlockIndex, widthBlockIndex, reconstructedFrame, blockSize)
-verticalRefernce(1:1,1:blockSize) = uint8(128);
-horizontalReference(1:blockSize,1:1) = uint8(128);
+verticalRefernce(1:1,1:blockSize) = int32(128);
+horizontalReference(1:blockSize,1:1) = int32(128);
 
 widthStart = int32((widthBlockIndex-1)*blockSize + 1);
 widthEnd = int32(widthStart + blockSize - 1);
